requirements.txt
simplematrixbotlib>=2.6.0,<3.0.0
pyyaml
aiohttp[speedups]
